+++ backend/ktnmgr/__init__.py
@@ -1,3 +1,3 @@
 """KTN Enclosure Manager."""
 
-__version__ = "1.2.10"
+__version__ = "1.3.0"